+++ web/apps/photos/src/services/plan.ts
@@ -1,6 +1,10 @@
 import { authenticatedRequestHeaders, ensureOk } from "@/base/http";
 import log from "@/base/log";
 import { apiURL, paymentsAppOrigin } from "@/base/origins";
+import {
+    getTotalFamilyUsage,
+    isPartOfFamily,
+} from "@/new/photos/services/user";
 import { nullToUndefined } from "@/utils/transform";
 import HTTPService from "@ente/shared/network/HTTPService";
 import {
@@ -10,6 +14,7 @@ import {
 } from "@ente/shared/storage/localStorage";
 import { getToken } from "@ente/shared/storage/localStorage/helpers";
 import isElectron from "is-electron";
+import { BonusData, UserDetails } from "types/user";
 import { z } from "zod";
 
 const PlanPeriod = z.enum(["month", "year"]);
@@ -223,3 +228,71 @@ export const redirectToCustomerPortal = async () => {
         .parse(await res.json()).data;
     window.location.href = data.url;
 };
+
+/**
+ * Return true if the given {@link Subscription} has not expired.
+ */
+export const isSubscriptionActive = (subscription: Subscription) =>
+    subscription && subscription.expiryTime > Date.now() * 1000;
+
+/**
+ * Return true if the given active {@link Subscription} is for a paid plan.
+ */
+export const isSubscriptionActivePaid = (subscription: Subscription) =>
+    subscription &&
+    isSubscriptionActive(subscription) &&
+    subscription.productID != "free";
+
+/**
+ * Return true if the given active {@link Subscription} is for a free plan.
+ */
+export const isSubscriptionActiveFree = (subscription: Subscription) =>
+    subscription &&
+    isSubscriptionActive(subscription) &&
+    subscription.productID == "free";
+
+/**
+ * Return true if the given {@link Subscription} is using Stripe.
+ */
+export const isSubscriptionStripe = (subscription: Subscription) =>
+    subscription && subscription.paymentProvider == "stripe";
+
+/**
+ * Return true if the given {@link Subscription} has the cancelled attribute.
+ */
+export const isSubscriptionCancelled = (subscription: Subscription) =>
+    subscription && subscription.attributes.isCancelled;
+
+export function isSubscriptionPastDue(subscription: Subscription) {
+    const thirtyDaysMicroseconds = 30 * 24 * 60 * 60 * 1000 * 1000;
+    const currentTime = Date.now() * 1000;
+    return (
+        !isSubscriptionCancelled(subscription) &&
+        subscription.expiryTime < currentTime &&
+        subscription.expiryTime >= currentTime - thirtyDaysMicroseconds
+    );
+}
+
+// Checks if the bonus data contain any bonus whose type starts with 'ADD_ON'
+export function hasAddOnBonus(bonusData?: BonusData) {
+    return (
+        bonusData &&
+        bonusData.storageBonuses &&
+        bonusData.storageBonuses.length > 0 &&
+        bonusData.storageBonuses.some((bonus) =>
+            bonus.type.startsWith("ADD_ON"),
+        )
+    );
+}
+
+export function hasExceededStorageQuota(userDetails: UserDetails) {
+    const bonusStorage = userDetails.storageBonus ?? 0;
+    if (isPartOfFamily(userDetails.familyData)) {
+        const usage = getTotalFamilyUsage(userDetails.familyData);
+        return usage > userDetails.familyData.storage + bonusStorage;
+    } else {
+        return (
+            userDetails.usage > userDetails.subscription.storage + bonusStorage
+        );
+    }
+}
