+++ web/apps/photos/src/utils/ui/index.tsx
@@ -2,7 +2,7 @@ import { DialogBoxAttributes } from "@ente/shared/components/DialogBox/types";
 import InfoOutlined from "@mui/icons-material/InfoRounded";
 import { t } from "i18next";
 import { Trans } from "react-i18next";
-import { Subscription } from "services/billingService";
+import { Subscription } from "services/plan";
 
 export const getTrashFilesMessage = (
     deleteFileHelper,
